Fix formatting floating point variables with guile-2.0
When using guile-2.0 the current locale is brought in from the
environment and that includes LC_NUMERIC which affects how decimal
numbers are represented.

Reset LC_NUMERIC to make sure that the numbers are formatted with
a period (and not a comma) as a decimal point separator.

For instance this fixes the numbers in the postscript and svg output,
and in the command lines calling ghostscript.